% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{create_2d_color_legend}
\alias{create_2d_color_legend}
\title{Create 2D color legend plot for gene/ADT coexpression}
\usage{
create_2d_color_legend(input, myso)
}
\arguments{
\item{input}{Shiny reactiveValues object containing user inputs}

\item{myso}{a Seurat object}
}
\value{
legend for app coexpression plot
}
\description{
Create 2D color legend plot for gene/ADT coexpression
}
