% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_filehandling.R
\name{create_feather_files}
\alias{create_feather_files}
\title{Generate Feather files from data in a Seurat object.}
\usage{
create_feather_files(
  seurat_object,
  feather_dir = ".citeviz",
  feather_file_prefix = "sample"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing metadata, assays, and reductions for a CITE-seq experiment.}

\item{feather_dir}{A string specifying the directory in which to store feather files generated by this function. Do not include slashes. Default value is a hidden folder named ".citeviz" in the current working directory.}

\item{feather_file_prefix}{A string specifying the prefix for each feather file generated from an input RDS file.}
}
\value{
The input feather directory (feather_dir), invisibly.
}
\description{
This function is called before running the app so that the user's Seurat object can be converted into Feather files.
}
\examples{
\dontrun{
myso <- readRDS("path/to/RDS_file.rds")
create_feather_files(seurat_object = myso, feather_dir = ".citeviz", feather_file_prefix = "project123")
}
}
