% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{create_gating_df}
\alias{create_gating_df}
\title{Create gating dataframe}
\usage{
create_gating_df()
}
\value{
Empty gate dataframe for app purposes
}
\description{
Create gating dataframe
}
\examples{

# Takes no arguments
example_df <- create_gating_df()

}
