% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_datatable.R
\name{create_gating_dt}
\alias{create_gating_dt}
\title{Gating datatable for display in UI}
\usage{
create_gating_dt(temp_reactive_gating_df)
}
\arguments{
\item{temp_reactive_gating_df}{reactive gating dataframe to use as contents}
}
\value{
an interactive data table
}
\description{
Gating datatable for display in UI
}
