% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandling.R
\name{find_reduction_in_altSCE}
\alias{find_reduction_in_altSCE}
\title{Find which alternate experiment within a SingleCellExperiment object a given reduction can be found in. This is needed if a user wants to retrieve and plot the cell embeddings for a given reduction from a SingleCellExperiment object.}
\usage{
find_reduction_in_altSCE(alt_exp_name, sce_object, reduction_name)
}
\arguments{
\item{alt_exp_name}{A string specifying the name of an alternate experiment from a SingleCellExperiment object in which to search if a given reduction is present. The names of alternate experiments can be obtained by looking through SingleCellExperiment::altExpNames(x), where x is a SingleCellExperiment object.}

\item{sce_object}{A SingleCellExperiment object containing data from a CITE-seq experiment processed with Seurat.}

\item{reduction_name}{A string specifying the name of a reduction (e.g., "UMAP", "PCA", etc.).}
}
\value{
A string specifying the name of an alternate experiment (e.g., "ADT", "RNA", "SCT", etc.) in which the cell embeddings for a given reduction can be found.
}
\description{
Find which alternate experiment within a SingleCellExperiment object a given reduction can be found in. This is needed if a user wants to retrieve and plot the cell embeddings for a given reduction from a SingleCellExperiment object.
}
\examples{
\dontrun{
sce_object <- readRDS("path/to/RDS_file_containing_SCE_object.rds")
alt_exp_name <- SingleCellExperiment::altExpNames(sce_object)[1]
reduction_name <- "UMAP"
alt_exp_of_desired_reduction <- find_reduction_in_altSCE(alt_exp_name, sce_object, reduction_name)
}
}
