% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_class.R
\docType{class}
\name{Gate-class}
\alias{Gate-class}
\title{Gate class for gate objects}
\value{
a gate class object
}
\description{
The Gate class is the core data structure of CITEViz to store
gating information. Gates are initiated upon the first click of the "Gate"
button and turned into a reactive value to facilitate interactivity with 
R-Shiny. The Gate class holds the following information:

* counter <integer>: keep track of gate number
* assay_name <character>: keep track of which assay (RNA, ADT, etc) the gate data is coming from
* input_cells <list>: a list of cell barcodes going into the gate 
e.g. if a gate isn't initialized yet, this variable contains all the cells.
* input_coords <data.frame>: the coordinates of the input cells
* subset_cells <list>: a list of output cells
* subset_coords <data.frame>: the coordinates of the output cells
* x_axis <character>: the feature used to plot the x-axis
* y_axis <character>: the feature used to plot the y-axis
* gate_coords <list>: the coordinates of the gate boundaries drawn by plotly
* name_subset_cells <character>: the label of the set of output cells inputted by the user
* num_input_cells <integer>: number of input cells
* num_subset_cells <integer>: number of output cells
* total_num_cells_in_sample <integer>: total number of cell in a sample
* pct_subset_from_previous <numeric>: proportion of output cell WRT previous gate
* pct_subset_from_total <numeric>: proportion of output cell WRT total cells
}
\section{Slots}{

\describe{
\item{\code{counter}}{integer.}

\item{\code{assay_name}}{character.}

\item{\code{input_cells}}{list.}

\item{\code{input_coords}}{data.frame.}

\item{\code{subset_cells}}{list.}

\item{\code{subset_coords}}{data.frame.}

\item{\code{x_axis}}{character.}

\item{\code{y_axis}}{character.}

\item{\code{gate_coords}}{list.}

\item{\code{name_subset_cells}}{character.}

\item{\code{num_input_cells}}{integer.}

\item{\code{num_subset_cells}}{integer.}

\item{\code{total_num_cells_in_sample}}{integer.}

\item{\code{pct_subset_from_previous}}{numeric.}

\item{\code{pct_subset_from_total}}{numeric.}
}}

