% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_class.R
\name{Gate}
\alias{Gate}
\title{Constructor for Gate objects}
\usage{
Gate(
  counter = NA_integer_,
  assay_name = NA_character_,
  input_cells = list(),
  input_coords = data.frame(),
  subset_cells = list(),
  subset_coords = data.frame(),
  x_axis = NA_character_,
  y_axis = NA_character_,
  gate_coords = list(),
  name_subset_cells = NA_character_,
  num_input_cells = NA_integer_,
  num_subset_cells = NA_integer_,
  total_num_cells_in_sample = NA_integer_,
  pct_subset_from_previous = NA_real_,
  pct_subset_from_total = NA_real_
)
}
\arguments{
\item{counter}{integer to keep track of gate number}

\item{assay_name}{character. Seurat object assay gates on}

\item{input_cells}{list. cells plotted}

\item{input_coords}{dataframe of plotted cell coordinates}

\item{subset_cells}{list. cells selected}

\item{subset_coords}{dataframe of selected cell coordinates}

\item{x_axis}{character. Feature on x-axis}

\item{y_axis}{character. Feature on y-axis}

\item{gate_coords}{list. coordinates of drawn gate}

\item{name_subset_cells}{character. user input of selected cell subset name}

\item{num_input_cells}{integer. number of total cells on plot}

\item{num_subset_cells}{integer. number of cells selected}

\item{total_num_cells_in_sample}{integer. number of total cells in seurat object}

\item{pct_subset_from_previous}{numeric. percentage of cells selected from cells plotted}

\item{pct_subset_from_total}{numeric. percentage of cells selected from cells in Seurat object}
}
\value{
Gate class object
}
\description{
All parameters are slots within the gate object
}
