% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{get_bilinear_val}
\alias{get_bilinear_val}
\title{Bilinear Interpolation}
\usage{
get_bilinear_val(x, y, ngrid, quad11, quad21, quad12, quad22)
}
\arguments{
\item{x}{integer vector}

\item{y}{integer vector}

\item{ngrid}{integer taken from get_color_matrix_df}

\item{quad11}{integer}

\item{quad21}{integer}

\item{quad12}{integer}

\item{quad22}{integer}
}
\value{
numeric vector for color values used in coexpression plot
}
\description{
Takes in a given x and y coordinate, the dimension of a square grid, and four values representing the base red, green, or blue (RGB) color values (0-255) of the four quadrants of the square grid
Returns the bilinear interpolated red, green, or blue value (0-255) of the given input coordinates (x,y)
}
