% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandling.R
\name{get_choices}
\alias{get_choices}
\title{Get dropdown menu options for selectInput elements in CITEViz.}
\usage{
get_choices(
  category,
  input_data_type,
  rds_object,
  arrow_file_list,
  input_file_df,
  assay_name = NULL
)
}
\arguments{
\item{category}{A string specifying the category of data to retrieve. Categories are based on slots in a Seurat object produced from a CITE-seq experiment Possible values are "metadata", "reductions", or "assays".}

\item{input_data_type}{An integer value indicating if the user-uploaded input files are RDS files holding Seurat objects or SingleCellExperiment objects, or Feather/Arrow files (1 = Seurat object, 2 = SingleCellExperiment object, 3 = Feather/Arrow file, etc.)}

\item{rds_object}{An RDS object containing metadata, assays, and reductions for a CITE-seq experiment. Can be NULL if the user uploaded Feather/Arrow files instead of an RDS file.}

\item{arrow_file_list}{A vector containing a list of Feather/Arrow files generated by CITEViz::create_feather_files() and uploaded by the user in the CITEViz UI. Can be NULL if the user uploaded an RDS file instead of Feather/Arrow files.}

\item{input_file_df}{A dataframe generated by retrieving file input data from a fileInput element in the UI (i.e., by running input$file_input in the app server). This dataframe contains information about each file uploaded by the user, such as the name of each file as it appears on the user's local filesystem (input_file_df$name) and the temporary datapath with which data from the file can be read in (input_file_df$datapath).}

\item{assay_name}{A string specifying the name of an assay (e.g., "ADT", "RNA", etc.) from a Seurat-processed CITE-seq experiment.}
}
\value{
A character vector of items with which to populate a dropdown menu in the CITEViz UI, sorted in ascending order.
}
\description{
This function retrieves various categories of data (e.g., metadata, reductions, or assays) from user-uploaded input file(s) (e.g., an RDS file containing a Seurat object or Feather files) and returns a character vector of items with which to populate a dropdown menu in the CITEViz user interface. This function can also get subchoices for a dropdown menu after a user has selected an assay they want to view. For example, if a user selects the "ADT" assay, then this function will return a vector of all the possible ADTs a user can choose to view from their input data. If a user selects the "RNA" assay, then this function will return a vector of all the genes a user can choose to view from their input data.
}
\examples{
\dontrun{
# if RDS file is uploaded by user
myso <- readRDS("path/to/RDS_file.rds")
input_file_data <- input$file_input

get_choices("metadata", 
input_data_type = FALSE, 
rds_object = myso, 
arrow_file_list = NULL, 
input_file_df = input_file_data, 
assay_name = NULL)

# if Feather/Arrow files are uploaded by user
input_file_data <- input$file_input
arrow_filenames <- input_file_data$name

get_choices("metadata", 
input_data_type = TRUE, 
rds_object = NULL, 
arrow_file_list = arrow_filenames, 
input_file_df = input_file_data, 
assay_name = NULL)
}

}
