% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandling.R
\name{get_data}
\alias{get_data}
\title{Title}
\usage{
get_data(
  category,
  input_data_type,
  rds_object,
  arrow_file_list,
  input_file_df,
  assay_name = NULL,
  reduction_name = NULL,
  assay_data_to_get = NULL
)
}
\arguments{
\item{category}{A string specifying the category of data to retrieve. Categories are based on slots in a Seurat object produced from a CITE-seq experiment Possible values are "metadata", "reductions", or "assays".}

\item{input_data_type}{An integer value indicating if the user-uploaded input files are RDS files holding Seurat objects or SingleCellExperiment objects, or Feather/Arrow files (1 = Seurat object, 2 = SingleCellExperiment object, 3 = Feather/Arrow file, etc.)}

\item{rds_object}{An RDS object containing metadata, assays, and reductions for a CITE-seq experiment. Can be NULL if the user uploaded Feather/Arrow files instead of an RDS file.}

\item{arrow_file_list}{A vector containing a list of Feather/Arrow files generated by CITEViz::create_feather_files() and uploaded by the user in the CITEViz UI. Can be NULL if the user uploaded an RDS file instead of Feather/Arrow files.}

\item{input_file_df}{A dataframe generated by retrieving file input data from a fileInput element in the UI (i.e., by running input$file_input in the app server). This dataframe contains information about each file uploaded by the user, such as the name of each file as it appears on the user's local filesystem (input_file_df$name) and the temporary datapath with which data from the file can be read in (input_file_df$datapath).}

\item{assay_name}{A string specifying the name of an assay (e.g., "ADT", "RNA", etc.) from a Seurat-processed CITE-seq experiment. Can be NULL if not retrieving any assay data.}

\item{reduction_name}{A string specifying the name of an reduction (e.g., "pca", "umap", etc.) from a Seurat-processed CITE-seq experiment. Can be NULL if not retrieving any reduction data.}

\item{assay_data_to_get}{Character vector specifying the column(s) from assay data to get (i.e. if the assay is "ADT", then a possible value for assay_data_to_get may be c("ab-CD10", "ab-CD38")). Can be NULL if not retrieving any assay data.}
}
\value{
A dataframe containing metadata, assay count data, or reduction embeddings data that was generated in a Seurat-processed CITE-seq experiment. The rownames of the dataframe are cell barcodes, and the column names are metadata columns, assay columns (i.e. if assay is "RNA", then assay columns would be genes), or reduction embedding columns (i.e. if the reduction is "pca", the embedding data colummns would be PC1, PC2, etc).
}
\description{
Title
}
\examples{
\dontrun{
# if RDS file is uploaded by user
myso <- readRDS("path/to/RDS_file.rds")
input_file_data <- input$file_input
get_data("reductions", input_data_type = FALSE, rds_object = myso, arrow_file_list = NULL, input_file_df = input_file_data, reduction_name = "pca")

# if Feather/Arrow files are uploaded by user
input_file_data <- input$file_input
arrow_filenames <- input_file_data$name
get_data("reductions", input_data_type = TRUE, rds_object = NULL, arrow_file_list = arrow_filenames, input_file_df = input_file_data, reduction_name = "pca")
}

}
