% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{get_palette}
\alias{get_palette}
\title{Custom color palette with color interpolation to use for QA plots}
\usage{
get_palette(num_colors = integer(3))
}
\arguments{
\item{num_colors}{number of colors}
}
\value{
a color palette
}
\description{
create custom color palette with color interpolation to use for QA plots
interpolate the base color palette so that exact number of colors in custom palette is same as number of unique values for selected metadata category
takes in an integer argument for number of colors to generate in palette
returns a character vector of color hex codes for the desired number of colors
}
