% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{set_gates_to_null}
\alias{set_gates_to_null}
\title{Reset all values in gate_reactive_values to NULL}
\usage{
set_gates_to_null(gate_name_string, local_gate_reactive_values)
}
\arguments{
\item{gate_name_string}{name of gate to be deleted}

\item{local_gate_reactive_values}{Shiny reactiveValues}
}
\value{
empty reactiveValues list of Gate objects
}
\description{
Reset all values in gate_reactive_values to NULL
}
\examples{

\dontrun{
   # Suppose reactive_list_of_gates has two gates: "gate1" and "gate2"
   reactive_list_of_gates <- set_gates_to_null("gate1", reactive_list_of_gates)
   reactive_list_of_gates <- set_gates_to_null("gate2", reactive_list_of_gates)
   
   # reactive_list_of_gates is now empty
}

}
