% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_class.R
\name{SetSubsetName}
\alias{SetSubsetName}
\title{Set Cell Subset Name}
\usage{
SetSubsetName(gate_obj, new_name)
}
\arguments{
\item{gate_obj}{a single Gate class object}

\item{new_name}{character. New name of cell subset}
}
\value{
Gate object with new name of subset cells in name_subset_cells slot
}
\description{
Set Cell Subset Name
}
