% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{update_gating_df}
\alias{update_gating_df}
\title{Populate/update gating data frame}
\usage{
update_gating_df(gate_name_string, reactive_gate_list, temp_gating_df)
}
\arguments{
\item{gate_name_string}{character string of new gate name}

\item{reactive_gate_list}{shiny reactivelist of gates}

\item{temp_gating_df}{empty gating dataframe built from create_gating_df function}
}
\value{
updated gating dataframe
}
\description{
Populate/update gating data frame
}
