% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_filehandling.R
\name{write_seurat_to_feather}
\alias{write_seurat_to_feather}
\title{Generate a Feather file from data in a Seurat object.}
\usage{
write_seurat_to_feather(
  assay_or_reduction_name = NULL,
  seurat_object,
  slot_name,
  feather_dir = ".citeviz",
  feather_file_prefix = "sample"
)
}
\arguments{
\item{assay_or_reduction_name}{A string specifying the name of an assay (e.g., ADT, RNA, etc.) or reduction (e.g., PCA, UMAP, etc.) present in the Seurat object that is passed as a parameter to this function.}

\item{seurat_object}{A Seurat object containing metadata, assays, and reductions for a CITE-seq experiment.}

\item{slot_name}{A string specifying one of three possible values: "metadata", "assays", or "reductions".}

\item{feather_dir}{A string specifying the directory in which to store feather files generated by this function. Do not include slashes. Default value is a hidden folder named ".citeviz" in the current working directory.}

\item{feather_file_prefix}{A string specifying the prefix for each feather file generated from an input RDS file.}
}
\value{
The input feather directory (feather_dir), invisibly.
}
\description{
This function retrieves data in specified slots ("metadata", "assays", or "reductions") from a user-inputted Seurat object and writes that data to a Feather file.
}
\examples{
\dontrun{
myso <- readRDS("path/to/RDS_file.rds")
write_seurat_to_feather(seurat_object = myso, slot_name = "metadata", feather_dir = ."citeviz", feather_file_prefix = "project123")
}
}
