% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{create_2d_color_legend}
\alias{create_2d_color_legend}
\title{Create 2D color legend plot for gene/ADT coexpression}
\usage{
create_2d_color_legend(input, myso)
}
\arguments{
\item{input}{Shiny internal parameter object containing UI user input values}

\item{myso}{a Seurat object}
}
\value{
legend for app coexpression plot
}
\description{
Create 2D color legend plot for gene/ADT coexpression
}
\examples{
\dontrun{
seurat_object <- readRDS("path/to/RDSfile/containing/Seurat/object")
create_2d_color_legend(input, seurat_object)
}
}
