% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandling.R
\name{create_feather_files}
\alias{create_feather_files}
\title{Generate Feather files from data in a Seurat object.}
\usage{
create_feather_files(
  seurat_object,
  feather_parent_dir = "citeviz_arrow",
  feather_file_prefix = "project1"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing metadata, assays, and reductions for a CITE-seq experiment.}

\item{feather_parent_dir}{A string specifying the directory in which all project-specific Feather folders/files are stored for this app. Do not include slashes. Default value is a folder named "citeviz_arrow" in the current working directory.}

\item{feather_file_prefix}{A string specifying the prefix for each feather file generated from an input Seurat object. These Feather files will be stored in a sub-directory named with the same prefix, and this subfolder can be found within the feather parent directory. It is recommended to set the feather file prefix to be the same as the base filename of the original RDS file the Seurat object originated from (i.e., if the input Seurat object came from an RDS file called "my_project_data.rds", then the prefix for the arrow files should be "my_project_data".)}
}
\value{
The path to the sub-directory (in the format "feather_parent_dir/feather_file_prefix") where Feather files generated from the same input Seurat object are stored. This path is relative to the user's current working directory. This value is returned invisibly.
}
\description{
This function is called before running the app so that the user's Seurat object can be converted into Feather files.
}
\examples{
\dontrun{
myso <- readRDS("path/to/RDS_file.rds")

create_feather_files(seurat_object = myso, 
feather_parent_dir = "citeviz_arrow", 
feather_file_prefix = "project123")
}
}
