% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_class.R
\docType{class}
\name{Gate-class}
\alias{Gate-class}
\title{Gate class for gate objects}
\value{
a gate class object
}
\description{
Gate class for gate objects
}
\section{Slots}{

\describe{
\item{\code{counter}}{integer.}

\item{\code{assay_name}}{character.}

\item{\code{input_cells}}{list.}

\item{\code{input_coords}}{data.frame.}

\item{\code{subset_cells}}{list.}

\item{\code{subset_coords}}{data.frame.}

\item{\code{x_axis}}{character.}

\item{\code{y_axis}}{character.}

\item{\code{gate_coords}}{list.}

\item{\code{name_subset_cells}}{character.}

\item{\code{num_input_cells}}{integer.}

\item{\code{num_subset_cells}}{integer.}

\item{\code{total_num_cells_in_sample}}{integer.}

\item{\code{pct_subset_from_previous}}{numeric.}

\item{\code{pct_subset_from_total}}{numeric.}
}}

