% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{get_color_matrix_df}
\alias{get_color_matrix_df}
\title{Create 2D color dataframe for gene/ADT coexpression}
\usage{
get_color_matrix_df(ngrid = 16)
}
\arguments{
\item{ngrid}{integer setting the resolution (dimensions) of color 
grid (e.g., ngrid = 16 sets a 16x16 color grid). Default is ngrid=16.}
}
\value{
dataframe for use in coexpression legend
}
\description{
Create 2D color dataframe for gene/ADT coexpression
}
