% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{get_reactive_gate_list}
\alias{get_reactive_gate_list}
\title{Convert Shiny reactiveValues object to reactive gate list}
\usage{
get_reactive_gate_list(gating_reactiveValues)
}
\arguments{
\item{gating_reactiveValues}{Shiny reactiveValues object holding all reactive gate objects}
}
\value{
reactive gate list for app purposes
}
\description{
Convert Shiny reactiveValues object to reactive gate list
}
