% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_class.R
\name{GetData}
\alias{GetData}
\title{Slot Accessor Method}
\usage{
GetData(gate_obj, slot_name)
}
\arguments{
\item{gate_obj}{a single Gate class object}

\item{slot_name}{name of slot to draw data from}
}
\value{
data from slot in gate object
}
\description{
Slot Accessor Method
}
