% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandling.R
\name{write_seurat_to_feather}
\alias{write_seurat_to_feather}
\title{Generate a Feather file from data in a Seurat object.}
\usage{
write_seurat_to_feather(
  assay_or_reduction_name = NULL,
  seurat_object,
  slot_name,
  output_dir_path,
  feather_file_prefix
)
}
\arguments{
\item{assay_or_reduction_name}{A string specifying the name of an assay (e.g., ADT, RNA, etc.) or reduction (e.g., PCA, UMAP, etc.) present in the Seurat object that is passed as a parameter to this function.}

\item{seurat_object}{A Seurat object containing metadata, assays, and reductions for a CITE-seq experiment.}

\item{slot_name}{A string specifying one of three possible values: "metadata", "assays", or "reductions".}

\item{output_dir_path}{A string specifying the relative path to an output directory in which to store the Feather file generated by this function. Feather files generated from the same input Seurat object should be stored in the same directory. Path is relative to the current working directory.}

\item{feather_file_prefix}{A string specifying the prefix for each Feather file generated from an input Seurat object.}
}
\value{
The relative path to the feather file that was generated. This value is returned invisibly.
}
\description{
This function retrieves data in specified slots ("metadata", "assays", or "reductions") from a user-inputted Seurat object and writes that data to a Feather file. The first column of data contains cell barcodes from a CITE-seq experiment, and subsequent columns contain metadata or count data for various assays (counts for each ADT, gene, etc.) and reductions (embeddings for UMAPs, PCAs, etc.). This columnar data storage format allows for much faster read access in downstream app processes.
}
\examples{
\dontrun{
myso <- readRDS("path/to/RDS_file.rds")

write_seurat_to_feather(seurat_object = myso, 
slot_name = "metadata", 
feather_parent_dir = "citeviz_arrow", 
feather_file_prefix = "project123")
}
}
